# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: AniMeat API
  description: |
    Server-authoritative API for the AniMeat food delivery platform.
    Source of truth for money, orders, payments, roles, and state transitions.
    All timestamps are RFC3339 UTC. Money is transmitted as integer paise (minor units)
    where a `*_paise` field exists; legacy rupee fields may still be present for
    backwards compatibility with legacy Flutter clients.
  version: 1.0.0
  contact:
    name: AniMeat Engineering
servers:
  - url: https://api.animeat.app
    description: Production
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /api/v1/orders:
    post:
      summary: Create order (server-authoritative pricing)
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
  /api/v1/orders/{id}:
    get:
      summary: Get a single order (visibility enforced server-side)
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
    post:
      summary: Cancel an order (customer only, pending states)
      operationId: cancelOrder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  order_id: { type: string }
        '400': { $ref: '#/components/responses/Error' }
  /api/v1/orders/{id}/transition:
    post:
      summary: Apply a role-authorized status transition
      operationId: transitionOrder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, preparing, ready_for_pickup, accepted, picked_up, delivered, cancelled]
      responses:
        '200':
          description: Updated order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '409': { $ref: '#/components/responses/Error' }
  /api/v1/payments/create-intent:
    post:
      summary: Create a payment intent/link for an order
      operationId: createPaymentIntent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order_id]
              properties:
                order_id: { type: string }
      responses:
        '200':
          description: Payment intent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentIntent'
...
  /api/v1/users/me:
    get:
      summary: Get own profile
      operationId: getProfile
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404': { $ref: '#/components/responses/Error' }
    put:
      summary: Upsert profile (created on first login)
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /api/v1/riders/me/availability:
    put:
      summary: Toggle online/offline
      operationId: setAvailability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_online]
              properties:
                is_online: { type: boolean }
      responses:
        '200':
          description: Rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
  /api/v1/riders/me/location:
    put:
      summary: Update GPS location
      operationId: updateRiderLocation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latitude, longitude]
              properties:
                latitude: { type: number }
                longitude: { type: number }
      responses:
        '200':
          description: Rider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rider'
...
  /api/v1/riders/orders/{id}/accept:
    post:
      summary: Claim an order
      operationId: acceptOrder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Assigned order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error, code, message]
      properties:
        error: { type: string }
        code: { type: string }
        message: { type: string }
        detail: { type: string }
...
    Order:
      type: object
      properties:
        id: { type: string }
        user_id: { type: string }
        vendor_id: { type: string, nullable: true }
        rider_id: { type: string, nullable: true }
        order_type: { type: string }
        status: { type: string }
        items: { type: array, items: { type: object } }
        subtotal_paise: { type: integer }
        delivery_fee_paise: { type: integer }
        discount_paise: { type: integer }
        total_amount_paise: { type: integer }
        total_amount: { type: number }
        payment_method: { type: string }
        payment_status: { type: string }
        created_at: { type: string, format: date-time }
    CreateOrderRequest:
      type: object
      required: [items]
      properties:
        vendor_id: { type: string }
        order_type: { type: string, enum: [meat, lunch] }
        items:
          type: array
          items:
            type: object
            required: [item_id, quantity]
            properties:
              item_id: { type: string }
              quantity: { type: integer }
        coupon_code: { type: string }
        payment_method: { type: string, enum: [cod, online] }
        delivery_street: { type: string }
        delivery_city: { type: string }
        delivery_zip: { type: string }
        delivery_lat: { type: number }
        delivery_lng: { type: number }
        special_notes: { type: string }
        idempotency_key: { type: string }
    PaymentIntent:
      type: object
      properties:
        order_id: { type: string }
        razorpay_order_id: { type: string }
        amount_paise: { type: integer }
        currency: { type: string }
        key_id: { type: string }
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: string }
        avatar_url: { type: string }
    UpdateProfileRequest:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        address: { type: string }
        avatar_url: { type: string }
...
    Rider:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        phone: { type: string }
        is_online: { type: boolean }
        is_approved: { type: boolean }
        approval_status: { type: string }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
...